refactor(filecopier test utils): organise file copier test utils
re #37
