+++ src/__testUtils__/FileCopier/filePaths.ts
@@ -1,0 +1,5 @@
+import { tmpdir } from 'os';
+
+export const SAMPLE_TEXT_FILE = `${tmpdir()}/sample.txt`;
+export const FAKE_IDML_FILE = `${tmpdir()}/fake.idml`;
+export const FAKE_FOLDER = `${tmpdir()}/FakeFolder`;
